chore: Moved singleton.mts and time.mts to utility/.
