+++ packages/architectura/src/utility/_index.mts
@@ -1,0 +1,2 @@
+export * from './singleton.mjs';
+export * from './time.mjs';
