+++ packages/architectura/src/utility/singleton.mts
