+++ packages/architectura/src/utility/time.mts
